Fisiologia/python-serial-realtime-app/requirements.txt
numpy==2.3.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
pyqtgraph==0.13.7
pyserial==3.5
